chore: split components into separate package
