+++ apps/web/src/components/Composer/Actions/AccessSettings/BasicSettings.tsx
@@ -1,8 +1,8 @@
 import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
-import { Button } from '@components/UI/Button';
 import { Card } from '@components/UI/Card';
 import { CollectionIcon, UsersIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
+import { Button } from 'components/Button';
 import { CollectModules } from 'lens';
 import type { Dispatch, FC } from 'react';
 import toast from 'react-hot-toast';
