+++ apps/web/src/components/Home/EnableMessages.tsx
@@ -1,9 +1,9 @@
-import { Button } from '@components/UI/Button';
 import { Card } from '@components/UI/Card';
 import { MailIcon, MailOpenIcon } from '@heroicons/react/solid';
 import { Trans } from '@lingui/macro';
 import { Client } from '@xmtp/xmtp-js';
 import clsx from 'clsx';
+import { Button } from 'components/Button';
 import { XMTP_ENV } from 'data/constants';
 import { useRouter } from 'next/router';
 import type { FC } from 'react';
