+++ apps/web/src/components/Profile/FollowDialog.tsx
@@ -1,9 +1,9 @@
 import Follow from '@components/Shared/Follow';
 import Slug from '@components/Shared/Slug';
 import SuperFollow from '@components/Shared/SuperFollow';
-import { Button } from '@components/UI/Button';
 import { Image } from '@components/UI/Image';
 import { t, Trans } from '@lingui/macro';
+import { Button } from 'components/Button';
 import type { Profile } from 'lens';
 import type { Dispatch, FC } from 'react';
 import formatHandle from 'utils/formatHandle';
