+++ apps/web/src/components/Profile/NftGallery/Create.tsx
@@ -1,10 +1,10 @@
 import { useApolloClient } from '@apollo/client';
 import EmojiPicker from '@components/Shared/EmojiPicker';
-import { Button } from '@components/UI/Button';
 import { Modal } from '@components/UI/Modal';
 import { Spinner } from '@components/UI/Spinner';
 import { ChevronLeftIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
+import { Button } from 'components/Button';
 import { ERROR_MESSAGE } from 'data/constants';
 import type { NftGallery } from 'lens';
 import {
