+++ apps/web/src/components/Shared/SuperFollow/FollowModule.tsx
@@ -1,5 +1,4 @@
 import AllowanceButton from '@components/Settings/Allowance/Button';
-import { Button } from '@components/UI/Button';
 import { Spinner } from '@components/UI/Spinner';
 import { WarningMessage } from '@components/UI/WarningMessage';
 import { StarIcon, UserIcon } from '@heroicons/react/outline';
@@ -8,6 +7,7 @@ import onError from '@lib/onError';
 import splitSignature from '@lib/splitSignature';
 import { t, Trans } from '@lingui/macro';
 import { LensHub } from 'abis';
+import { Button } from 'components/Button';
 import { LENSHUB_PROXY, POLYGONSCAN_URL, SIGN_WALLET } from 'data/constants';
 import type { ApprovedAllowanceAmount, Profile } from 'lens';
 import {
