+++ apps/web/src/components/Shared/SwitchNetwork.tsx
@@ -1,7 +1,7 @@
-import { Button } from '@components/UI/Button';
 import { SwitchHorizontalIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import { t, Trans } from '@lingui/macro';
+import { Button } from 'components/Button';
 import type { FC } from 'react';
 import toast from 'react-hot-toast';
 import { CHAIN_ID } from 'src/constants';
