+++ apps/web/src/components/UI/Alert.tsx
@@ -1,8 +1,7 @@
 import { Dialog, Transition } from '@headlessui/react';
+import { Button } from 'components/Button';
 import type { FC, ReactNode } from 'react';
 import { Fragment } from 'react';
-
-import { Button } from './Button';
 
 interface AlertProps {
   title: ReactNode;
