+++ apps/web/tailwind.config.js
@@ -3,7 +3,7 @@ const defaultTheme = require('tailwindcss/defaultTheme');
 const colors = require('tailwindcss/colors');
 
 module.exports = {
-  content: ['./src/**/*.{ts,tsx}'],
+  content: ['./src/**/*.{ts,tsx}','../../packages/components/**/*.{ts,tsx}'],
   darkMode: 'class',
   theme: {
     screens: {
